basic implementation of street tile + example json

+++ src/main/java/oogasalad/view/Coordinate.java
@@ -2,9 +2,9 @@ package oogasalad.view;
 
 public class Coordinate {
 
-  private int xCoor;
-  private int yCoor;
-  public Coordinate(int x, int y) {
+  private double xCoor;
+  private double yCoor;
+  public Coordinate(double x, double y) {
     //TODO: add error handling for bounds of board/tiles
     xCoor = x;
     yCoor = y;
@@ -18,9 +18,9 @@ public class Coordinate {
     yCoor = y;
   }
 
-  public int getXCoor() {return xCoor;}
+  public double getXCoor() {return xCoor;}
 
-  public int getYCoor() {return yCoor;}
+  public double getYCoor() {return yCoor;}
 
   @Override
   public boolean equals(Object o) {
